guestbook-project/ansible/roles/ec2_setup/files/app/backend/requirements.txt
Flask==2.3.2
boto3
psycopg2-binary
Pillow